+++ app/[courseCode]/addAssignments/page.jsx
@@ -1,143 +1,118 @@
-'use client';
-import { useState } from 'react'; // Import useState hook
-import CourseNavBar from '../../components/CourseNavBar';
+"use client";
+import { useState } from 'react';
 import Sidebar from '../../components/Sidebar';
-import db from '../../lib/firebase';
-import {auth,firestore,storage,uploadBytes} from '../../lib/firebase';
-import { addDoc,collection,updateDoc } from 'firebase/firestore';
-import {getDownloadURL, ref} from 'firebase/storage';
-import { update } from 'firebase/database';
 
 export default function Assignments() {
-    // State variables to store the title of the assignment, the selected type (quiz or essay), the due date, the worth of the assignment, and the PDF file
-    const [assignmentTitle, setAssignmentTitle] = useState('');
-    const [assignmentType, setAssignmentType] = useState('essay'); // Default type is 'essay'
-    const [dueDate, setDueDate] = useState('');
-    const [assignmentWorth, setAssignmentWorth] = useState('');
-    const [pdfFile, setPdfFile] = useState(null);
+    const [showForm, setShowForm] = useState(false);
+    const [questions, setQuestions] = useState([{ text: '', options: ['Option #1', 'Option #2'] }]);
+    const [weightage, setWeightage] = useState(0);
 
-    // Function to handle file upload
-    const handleFileUpload = (e) => {
-        const file = e.target.files[0];
-        setPdfFile(file);
+    const handleAddOption = (questionIndex) => {
+        setQuestions(questions.map((question, index) => {
+            if (index === questionIndex) {
+                return { ...question, options: [...question.options, `Option #${question.options.length + 1}`] };
+            }
+            return question;
+        }));
     };
 
-    // Function to handle submission of assignment details
-    const handleSubmit = async () => {
-        // Perform submission logic here
-        console.log('Assignment Details Submitted:', {
-            assignmentTitle,
-            assignmentType,
-            dueDate,
-            assignmentWorth,
-            pdfFile
-        });
-        // Clear input fields after submission
-        // setAssignmentTitle('');
-        // setDueDate('');
-        // setAssignmentWorth('');
-        // setPdfFile(null);
+    const handleDeleteQuestion = (questionIndex) => {
+        setQuestions(questions.filter((_, index) => index !== questionIndex));
+    };
 
-        try {
-            const colRef = collection(db,'Assignments');
+    const handleAddQuestion = () => {
+        setQuestions([...questions, { text: '', options: ['Option #1', 'Option #2'] }]);
+    };
 
-            const assignmentDocRef = await addDoc(colRef, {
-                assignmentTitle: assignmentTitle,
-                assignmentType: assignmentType,
-                dueDate: dueDate,
-                assignmentWorth: assignmentWorth,
-            });
+    const handleCorrectOptionChange = (questionIndex, optionIndex) => {
+        setQuestions(questions.map((question, index) => {
+            if (index === questionIndex) {
+                return { ...question, correctOption: optionIndex };
+            }
+            return question;
+        }));
+    };
 
-            let url = '';
-            const storageRef = ref(storage, `assignments/${assignmentDocRef.id}`);
-            
-            const snapshot = await uploadBytes(storageRef,pdfFile);
-
-            getDownloadURL(snapshot.ref).then((downloadURL) => {
-                updateDoc(assignmentDocRef,{fileURL:downloadURL});
-            })
-              
-    
-        } catch (error) {
-            //Handle errors if any thrown after validation
-            setErrorMsg(error.message);
+    const handleWeightageChange = (e) => {
+        const value = e.target.value;
+        if (value >= 0 && value <= 100) {
+            setWeightage(value);
+        } else {
+            alert('Weightage must be a number between 0 and 100.');
         }
-
-
     };
 
     return (
-        <div className="flex flex-col h-screen bg-blue-100"> {/* Add h-screen class to ensure full height */}
+        <div className="flex flex-col h-screen bg-blue-100 overflow-auto">
             <Sidebar />
             <div className="p-6 text-center w-full">
                 <h1 className="text-3xl text-black font-semibold mb-4" data-testid="course-heading">Course Name</h1>
                 <h2 className="text-3xl text-black font mt-4" data-testid="assignments-heading"> New Assignment</h2>
-                <div className="flex flex-col items-center mt-5"> {/* Center the content vertically */}
-                    {/* Radio buttons for selecting the assignment type */}
-                    <div className="flex items-center mb-4">
-                        <input
-                            type="radio"
-                            id="quiz"
-                            name="assignmentType"
-                            value="quiz"
-                            checked={assignmentType === 'quiz'}
-                            onChange={() => setAssignmentType('quiz')} // Update the assignment type state to 'quiz'
-                            className="mr-2 text-black"
-                        />
-                        <label htmlFor="quiz" className="mr-4 text-black">Quiz</label>
-                        <input
-                            type="radio"
-                            id="essay"
-                            name="assignmentType"
-                            value="essay"
-                            checked={assignmentType === 'essay'}
-                            onChange={() => setAssignmentType('essay')} // Update the assignment type state to 'essay'
-                            className="mr-2 text-black"
-                        />
-                        <label htmlFor="essay" className="text-black">Essay</label>
-                    </div>
-                    {/* Input field for the title of the assignment */}
-                    <input
-                        type="text"
-                        placeholder="Enter assignment title"
-                        className="border border-gray-300 px-4 py-2 rounded-md text-black w-96 mb-4" // Set text color to black and width to 96
-                        value={assignmentTitle}
-                        onChange={(e) => setAssignmentTitle(e.target.value)} // Update the assignment title state
-                    />
-                    {/* Input field for the due date of the assignment */}
-                    <input
-                        type="date"
-                        placeholder="Enter due date"
-                        className="border border-gray-300 px-4 py-2 rounded-md text-black w-96 mb-4" // Set text color to black and width to 96
-                        value={dueDate}
-                        onChange={(e) => setDueDate(e.target.value)} // Update the due date state
-                    />
-                    {/* Input field for the worth of the assignment */}
-                    <input
-                        type="number"
-                        placeholder="Enter assignment worth"
-                        className="border border-gray-300 px-4 py-2 rounded-md text-black w-96 mb-4" // Set text color to black and width to 96
-                        value={assignmentWorth}
-                        onChange={(e) => setAssignmentWorth(e.target.value)} // Update the assignment worth state
-                    />
-                    {/* Input field for uploading PDF file */}
-                    <input
-                        type="file"
-                        accept=".pdf"
-                        onChange={handleFileUpload} // Handle file upload
-                        className="border border-gray-300 px-4 py-2 rounded-md text-black w-96 mb-4" // Set text color to black and width to 96
-                    />
-                    {/* Button to submit the assignment details */}
-                    <button
-                        className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-                        onClick={handleSubmit} // Handle submission of assignment details
-                    >
-                        Submit
-                    </button>
+                <div className="flex flex-row items-center justify-center p-4">
+                    <button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mt-4 mx-5" onClick={() => setShowForm(true)}>Add Quiz</button>
+                    <a href='addEssays' className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mt-4 mx-5">Add Essay</a>
+                    <button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mt-4 mx-5" onClick={handleAddQuestion}>Add Question</button>
                 </div>
-                <div className="overflow-x-auto">
-                    {/* Additional content can be added here */}
-                </div>
+                {showForm && (
+                    <>
+                        <div className="flex flex-row items-center justify-center p-4 text-black font-semibold">
+                            <label className="mr-2">Weightage of the quiz in the course:</label>
+                            <input type="number" value={weightage} onChange={handleWeightageChange} min="0" max="100" />
+                        </div>
+                        {questions.map((question, questionIndex) => (
+                    
+                    <div key={questionIndex} className="mt-4">
+                        <input
+                            type="text"
+                            placeholder="Enter question"
+                            className="border border-gray-300 px-4 py-2 rounded-md text-black w-96 mb-4"
+                            value={question.text}
+                            onChange={(e) => setQuestions(questions.map((q, i) => {
+                                if (i === questionIndex) {
+                                    return { ...q, text: e.target.value };
+                                }
+                                return q;
+                            }))}
+                        />
+                        <button className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded mx-4" onClick={() => handleDeleteQuestion(questionIndex)}>Delete Question</button>
+                        <h3 className="text-lg text-black font-semibold mb-2">Select Correct Option</h3>
+                        {question.options.map((option, optionIndex) => (
+                            <div key={optionIndex} className="flex flex-row justify-center items-center mt-2">
+                                <input
+                                    type="radio"
+                                    className='mx-10'
+                                    checked={question.correctOption === optionIndex}
+                                    onChange={() => handleCorrectOptionChange(questionIndex, optionIndex)}
+                                />
+                                <input
+                                    type="text"
+                                    value={option}
+                                    className="border border-gray-300 px-4 py-2 rounded-md text-black w-64 mr-4"
+                                    onChange={(e) => setQuestions(questions.map((q, i) => {
+                                        if (i === questionIndex) {
+                                            return { ...q, options: q.options.map((o, j) => {
+                                                if (j === optionIndex) {
+                                                    return e.target.value;
+                                                }
+                                                return o;
+                                            }) };
+                                        }
+                                        return q;
+                                    }))}
+                                />
+                                <button className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded" onClick={() => setQuestions(questions.map((q, i) => {
+                                    if (i === questionIndex) {
+                                        return { ...q, options: q.options.filter((o, j) => j !== optionIndex) };
+                                    }
+                                    return q;
+                                }))}>Delete Option</button>
+                            </div>
+                        ))}
+                        <button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mt-2" onClick={() => handleAddOption(questionIndex)}>Add Option</button>
+                        </div>
+                        ))}
+                    </>
+                )}
             </div>
         </div>
     );
